// src/types/cameraTypes.ts
export enum CAMERA {
  ALL = "all",
  FHAZ = "FHAZ",
  RHAZ = "RHAZ",
  MAST = "MAST",
  CHEMCAM = "CHEMCAM",
  MAHLI = "MAHLI",
  MARDI = "MARDI",
  NAVCAM = "NAVCAM",
  PANCAM = "PANCAM",
  MINITES = "MINITES",
}

export enum PERSEVERENCE_CAMERAS {
  EDL_RUCAM = "EDL_RUCAM",
  EDL_DDCAM = "EDL_DDCAM",
  EDL_PUCAM1 = "EDL_PUCAM1",
  EDL_PUCAM2 = "EDL_PUCAM2",
  NAVCAM_LEFT = "NAVCAM_LEFT",
  NAVCAM_RIGHT = "NAVCAM_RIGHT",
  MCZ_RIGHT = "MCZ_RIGHT",
  MCZ_LEFT = "MCZ_LEFT",
  FRONT_HAZCAM_LEFT_A = "FRONT_HAZCAM_LEFT_A",
  FRONT_HAZCAM_RIGHT_A = "FRONT_HAZCAM_RIGHT_A",
  REAR_HAZCAM_LEFT = "REAR_HAZCAM_LEFT",
  REAR_HAZCAM_RIGHT = "REAR_HAZCAM_RIGHT",
  SKYCAM = "SKYCAM",
  SHERLOC_WATSON = "SHERLOC_WATSON",
}